# Implement reference counting in MPT