--- a/pkg/core/dao/dao.go
+++ b/pkg/core/dao/dao.go
@@ -510,16 +510,28 @@ func makeStateRootKey(height uint32) []byte {
 }
 
 // InitMPT initializes MPT at the given height.
-func (dao *Simple) InitMPT(height uint32) error {
+func (dao *Simple) InitMPT(height uint32, enableRefCount bool) error {
+	var gcKey = []byte{byte(storage.DataMPT), 1}
 	if height == 0 {
-		dao.MPT = mpt.NewTrie(nil, dao.Store)
-		return nil
+		dao.MPT = mpt.NewTrie(nil, enableRefCount, dao.Store)
+		var val byte
+		if enableRefCount {
+			val = 1
+		}
+		return dao.Store.Put(gcKey, []byte{val})
+	}
+	var hasRefCount bool
+	if v, err := dao.Store.Get(gcKey); err == nil {
+		hasRefCount = v[0] != 0
+	}
+	if hasRefCount != enableRefCount {
+		return fmt.Errorf("KeepOnlyLatestState setting mismatch: old=%v, new=%v", hasRefCount, enableRefCount)
 	}
 	r, err := dao.GetStateRoot(height)
 	if err != nil {
 		return err
 	}
-	dao.MPT = mpt.NewTrie(mpt.NewHashNode(r.Root), dao.Store)
+	dao.MPT = mpt.NewTrie(mpt.NewHashNode(r.Root), enableRefCount, dao.Store)
 	return nil
 }