--- a/pkg/core/mpt/trie.go
+++ b/pkg/core/mpt/trie.go
@@ -2,7 +2,9 @@ package mpt
 
 import (
 	"bytes"
+	"encoding/binary"
 	"errors"
+	"fmt"
 
 	"github.com/nspcc-dev/neo-go/pkg/core/storage"
 	"github.com/nspcc-dev/neo-go/pkg/io"
@@ -13,7 +15,15 @@ import (
 type Trie struct {
 	Store *storage.MemCachedStore
 
-	root Node
+	root            Node
+	refcountEnabled bool
+	refcount        map[util.Uint256]*cachedNode
+}
+
+type cachedNode struct {
+	bytes    []byte
+	initial  int32
+	refcount int32
 }
 
 // ErrNotFound is returned when requested trie item is missing.
@@ -22,7 +32,7 @@ var ErrNotFound = errors.New("item not found")
 // NewTrie returns new MPT trie. It accepts a MemCachedStore to decouple storage errors from logic errors
 // so that all storage errors are processed during `store.Persist()` at the caller.
 // This also has the benefit, that every `Put` can be considered an atomic operation.
-func NewTrie(root Node, store *storage.MemCachedStore) *Trie {
+func NewTrie(root Node, enableRefCount bool, store *storage.MemCachedStore) *Trie {
 	if root == nil {
 		root = new(HashNode)
 	}
@@ -30,6 +40,9 @@ func NewTrie(root Node, store *storage.MemCachedStore) *Trie {
 	return &Trie{
 		Store: store,
 		root:  root,
+
+		refcountEnabled: enableRefCount,
+		refcount:        make(map[util.Uint256]*cachedNode),
 	}
 }
 
@@ -107,12 +120,15 @@ func (t *Trie) Put(key, value []byte) error {
 func (t *Trie) putIntoLeaf(curr *LeafNode, path []byte, val Node) (Node, error) {
 	v := val.(*LeafNode)
 	if len(path) == 0 {
+		t.removeRef(curr.Hash(), curr.bytes)
+		t.addRef(val.Hash(), val.Bytes())
 		return v, nil
 	}
 
 	b := NewBranchNode()
-	b.Children[path[0]] = newSubTrie(path[1:], v)
+	b.Children[path[0]] = t.newSubTrie(path[1:], v, true)
 	b.Children[lastChild] = curr
+	t.addRef(b.Hash(), b.bytes)
 	return b, nil
 }
 
@@ -120,18 +136,21 @@ func (t *Trie) putIntoLeaf(curr *LeafNode, path []byte, val Node) (Node, error) 
 // It returns Node if curr needs to be replaced and error if any.
 func (t *Trie) putIntoBranch(curr *BranchNode, path []byte, val Node) (Node, error) {
 	i, path := splitPath(path)
+	t.removeRef(curr.Hash(), curr.bytes)
 	r, err := t.putIntoNode(curr.Children[i], path, val)
 	if err != nil {
 		return nil, err
 	}
 	curr.Children[i] = r
 	curr.invalidateCache()
+	t.addRef(curr.Hash(), curr.bytes)
 	return curr, nil
 }
 
 // putIntoExtension puts val to trie if current node is an Extension.
 // It returns Node if curr needs to be replaced and error if any.
 func (t *Trie) putIntoExtension(curr *ExtensionNode, path []byte, val Node) (Node, error) {
+	t.removeRef(curr.Hash(), curr.bytes)
 	if bytes.HasPrefix(path, curr.key) {
 		r, err := t.putIntoNode(curr.next, path[len(curr.key):], val)
 		if err != nil {
@@ -139,6 +158,7 @@ func (t *Trie) putIntoExtension(curr *ExtensionNode, path []byte, val Node) (Nod
 		}
 		curr.next = r
 		curr.invalidateCache()
+		t.addRef(curr.Hash(), curr.bytes)
 		return curr, nil
 	}
 
@@ -147,16 +167,19 @@ func (t *Trie) putIntoExtension(curr *ExtensionNode, path []byte, val Node) (Nod
 	keyTail := curr.key[lp:]
 	pathTail := path[lp:]
 
-	s1 := newSubTrie(keyTail[1:], curr.next)
+	s1 := t.newSubTrie(keyTail[1:], curr.next, false)
 	b := NewBranchNode()
 	b.Children[keyTail[0]] = s1
 
 	i, pathTail := splitPath(pathTail)
-	s2 := newSubTrie(pathTail, val)
+	s2 := t.newSubTrie(pathTail, val, true)
 	b.Children[i] = s2
 
+	t.addRef(b.Hash(), b.bytes)
 	if lp > 0 {
-		return NewExtensionNode(copySlice(pref), b), nil
+		e := NewExtensionNode(copySlice(pref), b)
+		t.addRef(e.Hash(), e.bytes)
+		return e, nil
 	}
 	return b, nil
 }
@@ -165,7 +188,8 @@ func (t *Trie) putIntoExtension(curr *ExtensionNode, path []byte, val Node) (Nod
 // It returns Node if curr needs to be replaced and error if any.
 func (t *Trie) putIntoHash(curr *HashNode, path []byte, val Node) (Node, error) {
 	if curr.IsEmpty() {
-		return newSubTrie(path, val), nil
+		hn := t.newSubTrie(path, val, true)
+		return hn, nil
 	}
 
 	result, err := t.getFromStore(curr.hash)
@@ -176,13 +200,20 @@ func (t *Trie) putIntoHash(curr *HashNode, path []byte, val Node) (Node, error) 
 }
 
 // newSubTrie create new trie containing node at provided path.
-func newSubTrie(path []byte, val Node) Node {
+func (t *Trie) newSubTrie(path []byte, val Node, newVal bool) Node {
+	if newVal {
+		t.addRef(val.Hash(), val.Bytes())
+	}
 	if len(path) == 0 {
 		return val
 	}
-	return NewExtensionNode(path, val)
+	e := NewExtensionNode(path, val)
+	t.addRef(e.Hash(), e.bytes)
+	return e
 }
 
+// putIntoNode puts val with provided path inside curr and returns updated node.
+// Reference counters are updated for both curr and returned value.
 func (t *Trie) putIntoNode(curr Node, path []byte, val Node) (Node, error) {
 	switch n := curr.(type) {
 	case *LeafNode:
@@ -212,10 +243,13 @@ func (t *Trie) Delete(key []byte) error {
 
 func (t *Trie) deleteFromBranch(b *BranchNode, path []byte) (Node, error) {
 	i, path := splitPath(path)
+	h := b.Hash()
+	bs := b.bytes
 	r, err := t.deleteFromNode(b.Children[i], path)
 	if err != nil {
 		return nil, err
 	}
+	t.removeRef(h, bs)
 	b.Children[i] = r
 	b.invalidateCache()
 	var count, index int
@@ -228,6 +262,7 @@ func (t *Trie) deleteFromBranch(b *BranchNode, path []byte) (Node, error) {
 	}
 	// count is >= 1 because branch node had at least 2 children before deletion.
 	if count > 1 {
+		t.addRef(b.Hash(), b.bytes)
 		return b, nil
 	}
 	c := b.Children[index]
@@ -241,24 +276,32 @@ func (t *Trie) deleteFromBranch(b *BranchNode, path []byte) (Node, error) {
 		}
 	}
 	if e, ok := c.(*ExtensionNode); ok {
+		t.removeRef(e.Hash(), e.bytes)
 		e.key = append([]byte{byte(index)}, e.key...)
 		e.invalidateCache()
+		t.addRef(e.Hash(), e.bytes)
 		return e, nil
 	}
 
-	return NewExtensionNode([]byte{byte(index)}, c), nil
+	e := NewExtensionNode([]byte{byte(index)}, c)
+	t.addRef(e.Hash(), e.bytes)
+	return e, nil
 }
 
 func (t *Trie) deleteFromExtension(n *ExtensionNode, path []byte) (Node, error) {
 	if !bytes.HasPrefix(path, n.key) {
 		return nil, ErrNotFound
 	}
+	h := n.Hash()
+	bs := n.bytes
 	r, err := t.deleteFromNode(n.next, path[len(n.key):])
 	if err != nil {
 		return nil, err
 	}
+	t.removeRef(h, bs)
 	switch nxt := r.(type) {
 	case *ExtensionNode:
+		t.removeRef(nxt.Hash(), nxt.bytes)
 		n.key = append(n.key, nxt.key...)
 		n.next = nxt.next
 	case *HashNode:
@@ -269,13 +312,17 @@ func (t *Trie) deleteFromExtension(n *ExtensionNode, path []byte) (Node, error) 
 		n.next = r
 	}
 	n.invalidateCache()
+	t.addRef(n.Hash(), n.bytes)
 	return n, nil
 }
 
+// deleteFromNode removes value with provided path from curr and returns an updated node.
+// Reference counters are updated for both curr and returned value.
 func (t *Trie) deleteFromNode(curr Node, path []byte) (Node, error) {
 	switch n := curr.(type) {
 	case *LeafNode:
 		if len(path) == 0 {
+			t.removeRef(curr.Hash(), curr.Bytes())
 			return new(HashNode), nil
 		}
 		return nil, ErrNotFound
@@ -314,32 +361,88 @@ func makeStorageKey(mptKey []byte) []byte {
 // new node to storage. Normally, flush should be called with every StateRoot persist, i.e.
 // after every block.
 func (t *Trie) Flush() {
-	t.flush(t.root)
-}
-
-func (t *Trie) flush(node Node) {
-	if node.IsFlushed() {
-		return
-	}
-	switch n := node.(type) {
-	case *BranchNode:
-		for i := range n.Children {
-			t.flush(n.Children[i])
+	for h, node := range t.refcount {
+		if node.refcount != 0 {
+			if node.bytes == nil {
+				panic("item not in trie")
+			}
+			if t.refcountEnabled {
+				node.initial = t.updateRefCount(h)
+				if node.initial == 0 {
+					delete(t.refcount, h)
+				}
+			} else if node.refcount > 0 {
+				_ = t.Store.Put(makeStorageKey(h.BytesBE()), node.bytes)
+			}
+			node.refcount = 0
+		} else {
+			delete(t.refcount, h)
 		}
-	case *ExtensionNode:
-		t.flush(n.next)
-	case *HashNode:
-		return
 	}
-	t.putToStore(node)
 }
 
-func (t *Trie) putToStore(n Node) {
-	if n.Type() == HashT {
-		panic("can't put hash node in trie")
+// updateRefCount should be called only when refcounting is enabled.
+func (t *Trie) updateRefCount(h util.Uint256) int32 {
+	if !t.refcountEnabled {
+		panic("`updateRefCount` is called, but GC is disabled")
 	}
-	_ = t.Store.Put(makeStorageKey(n.Hash().BytesBE()), n.Bytes()) // put in MemCached returns no errors
-	n.SetFlushed()
+	var data []byte
+	key := makeStorageKey(h.BytesBE())
+	node := t.refcount[h]
+	cnt := node.initial
+	if cnt == 0 {
+		// A newly created item which may be in store.
+		var err error
+		data, err = t.Store.Get(key)
+		if err == nil {
+			cnt = int32(binary.LittleEndian.Uint32(data[len(data)-4:]))
+		}
+	}
+	if len(data) == 0 {
+		data = append(node.bytes, 0, 0, 0, 0)
+	}
+	cnt += node.refcount
+	switch {
+	case cnt < 0:
+		// BUG: negative reference count
+		panic(fmt.Sprintf("negative reference count: %s new %d, upd %d", h.StringBE(), cnt, t.refcount[h]))
+	case cnt == 0:
+		_ = t.Store.Delete(key)
+	default:
+		binary.LittleEndian.PutUint32(data[len(data)-4:], uint32(cnt))
+		_ = t.Store.Put(key, data)
+	}
+	return cnt
+}
+
+func (t *Trie) addRef(h util.Uint256, bs []byte) {
+	node := t.refcount[h]
+	if node == nil {
+		t.refcount[h] = &cachedNode{
+			refcount: 1,
+			bytes:    bs,
+		}
+		return
+	}
+	node.refcount++
+	if node.bytes == nil {
+		node.bytes = bs
+	}
+}
+
+func (t *Trie) removeRef(h util.Uint256, bs []byte) {
+	node := t.refcount[h]
+	if node == nil {
+		t.refcount[h] = &cachedNode{
+			refcount: -1,
+			bytes:    bs,
+		}
+		return
+	}
+	node.refcount--
+	if node.bytes == nil {
+		node.bytes = bs
+	}
 }
 
 func (t *Trie) getFromStore(h util.Uint256) (Node, error) {
@@ -354,6 +457,15 @@ func (t *Trie) getFromStore(h util.Uint256) (Node, error) {
 	if r.Err != nil {
 		return nil, r.Err
 	}
+
+	if t.refcountEnabled {
+		data = data[:len(data)-4]
+		node := t.refcount[h]
+		if node != nil {
+			node.bytes = data
+			node.initial = int32(r.ReadU32LE())
+		}
+	}
 	n.Node.(flushedNode).setCache(data, h)
 	return n.Node, nil
 }
@@ -366,6 +478,7 @@ func (t *Trie) Collapse(depth int) {
 		panic("negative depth")
 	}
 	t.root = collapse(depth, t.root)
+	t.refcount = make(map[util.Uint256]*cachedNode)
 }
 
 func collapse(depth int, node Node) Node {