Results for a full testnet dump (5003730 blocks):
```
# Master-2.x
real    106m36,227s
user    192m32,855s
sys     14m45,739s
13G     .

# Refcount
real    121m10,446s
user    231m6,476s
sys     17m0,014s
9,0G    .

# Generation GC (note: GC hadn't been finished when restore was completed)
real    123m35,329s
user    284m21,860s
sys     22m32,090s
9,5G    .
```

First 2M mainnet blocks (will publish 2M-3M later). Note, generation GC starts every 200k blocks, so not all garbage was collected).
```
# Generation GC
real    46m31,497s
user    76m51,146s
sys     5m52,001s
12G     .

# Master-2.x
real    44m28,469s
user    74m43,983s
sys     5m20,414s
13G     .

# Refcount
real    44m22,258s
user    74m16,260s
sys     5m5,966s
11G     .
```

First 3M mainnet blocks (same note for generational GC):
```
# Master-2.x
real    644m50,778s
user    1223m8,365s
sys     69m23,930s
56G     chains

# Generational
real    698m50,998s
user    1019m11,077s
sys     85m33,651s
46G     chains

# Refcount
real    510m3,492s
user    1167m14,155s
sys     40m38,908s
28G     chains
```

This can be used for benchmark, though `VerifyBlocks: false` must be set manually. https://gist.github.com/fyrchik/df3733df056cdf9c0ab6b39dfc6d759d

UPDATE:
Results for optimized refcount version (boltdb was used, leveldb TBD):
```
# Refcount (stored at the end of node, no optimizations)
real    106m48,438s
user    151m50,943s
sys     6m56,261s
16G     chains

# Master
real    117m21,565s
user    145m42,361s
sys     10m26,123s
22G     chains

# Refcount + flush only changed
real    106m49,477s
user    150m45,168s
sys     6m54,792s
16G     chains

# Refcount + flush only changed + cache refcount of nodes read from store
real    107m5,384s
user    151m29,465s
sys     6m59,464s
16G     chains
```

UPDATE:
Results for leveldb+testnet with config switch
```
# Master, no GC
real    106m4,500s
user   193m19,865s
sys     14m41,085s
12G chains

# GC disabled in config
real   108m51,511s
user  199m24,728s
sys    14m56,655s
12G	chains

# GC enabled
real    119m48,806s
user   229m39,195s
sys    16m48,748s
8,3G	chains
```