@roman-khimov 2020-10-28
So, 31% better in size and 13% worse in time for testnet.