@fyrchik 2020-11-06
@roman-khimov time difference should be even less if we enable block and tx verification.