@fyrchik 2020-11-18
```
# With pointers, refcount off
real    108m18,157s
user    197m42,104s
sys     15m6,246s
12G     chains

# With pointers, refcount on
real    118m18,436s
user    226m17,145s
sys     17m5,012s
8,3G    chains

# No pointers, refcount off
real    114m21,650s
user    213m47,008s
sys     15m48,879s
12G     chains

# No pointers, refcount on
Failed with panic early, will investigate.
```

But when comparing 3 first there is a ~6-7% performance decrease going from implementation using pointers to non-pointers one.