@roman-khimov 2020-11-18
OK, it looks like pointers win here despite numerous microallocations.